:truck: api: move findPrivateCertificatesByUserId method from lib to src
