+++ api/tests/certification/results/integration/infrastructure/repositories/certificate-repository_test.js
@@ -1180,6 +1180,309 @@ describe('Integration | Infrastructure | Repository | Certification', function (
       ]);
     });
   });
+
+  describe('#findPrivateCertificatesByUserId', function () {
+    it('should return an empty list when the certificate does not exist', async function () {
+      // when
+      const result = await certificateRepository.findPrivateCertificatesByUserId({ userId: 123 });
+
+      // then
+      expect(result).to.deep.equal([]);
+    });
+
+    it('should return an empty list when the certificate has no assessment-result', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser().id;
+      const privateCertificateData = {
+        firstName: 'Sarah Michelle',
+        lastName: 'Gellar',
+        birthdate: '1977-04-14',
+        birthplace: 'Saint-Ouen',
+        isPublished: true,
+        isCancelled: false,
+        userId,
+        date: new Date('2020-01-01'),
+        verificationCode: 'P-SOMECODE',
+        maxReachableLevelOnCertificationDate: 5,
+        deliveredAt: new Date('2021-05-05'),
+        certificationCenter: 'Centre des poules bien dodues',
+        pixScore: 51,
+        commentForCandidate: 'Il aime beaucoup les mangues, et ça se voit !',
+      };
+      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+      const sessionId = databaseBuilder.factory.buildSession({
+        publishedAt: privateCertificateData.deliveredAt,
+        certificationCenter: privateCertificateData.certificationCenter,
+        certificationCenterId,
+      }).id;
+      const certificationCourseId = databaseBuilder.factory.buildCertificationCourse({
+        firstName: privateCertificateData.firstName,
+        lastName: privateCertificateData.lastName,
+        birthdate: privateCertificateData.birthdate,
+        birthplace: privateCertificateData.birthplace,
+        isPublished: privateCertificateData.isPublished,
+        isCancelled: privateCertificateData.isCancelled,
+        createdAt: privateCertificateData.date,
+        verificationCode: privateCertificateData.verificationCode,
+        maxReachableLevelOnCertificationDate: privateCertificateData.maxReachableLevelOnCertificationDate,
+        sessionId,
+        userId,
+      }).id;
+      databaseBuilder.factory.buildAssessment({ certificationCourseId });
+      await databaseBuilder.commit();
+
+      // when
+      const result = await certificateRepository.findPrivateCertificatesByUserId({ userId });
+
+      // then
+      expect(result).to.deep.equal([]);
+    });
+
+    it('should return the certificate when it is cancelled', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser().id;
+      const privateCertificateData = {
+        firstName: 'Sarah Michelle',
+        lastName: 'Gellar',
+        birthdate: '1977-04-14',
+        birthplace: 'Saint-Ouen',
+        isPublished: true,
+        isCancelled: true,
+        userId,
+        date: new Date('2020-01-01'),
+        verificationCode: 'P-SOMECODE',
+        maxReachableLevelOnCertificationDate: 5,
+        deliveredAt: new Date('2021-05-05'),
+        certificationCenter: 'Centre des poules bien dodues',
+        pixScore: 51,
+      };
+      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+      const sessionId = databaseBuilder.factory.buildSession({
+        publishedAt: privateCertificateData.deliveredAt,
+        certificationCenter: privateCertificateData.certificationCenter,
+        certificationCenterId,
+      }).id;
+      const certificationCourseId = databaseBuilder.factory.buildCertificationCourse({
+        firstName: privateCertificateData.firstName,
+        lastName: privateCertificateData.lastName,
+        birthdate: privateCertificateData.birthdate,
+        birthplace: privateCertificateData.birthplace,
+        isPublished: privateCertificateData.isPublished,
+        isCancelled: privateCertificateData.isCancelled,
+        createdAt: privateCertificateData.date,
+        verificationCode: privateCertificateData.verificationCode,
+        maxReachableLevelOnCertificationDate: privateCertificateData.maxReachableLevelOnCertificationDate,
+        sessionId,
+        userId,
+      }).id;
+      databaseBuilder.factory.buildAssessmentResult.last({
+        certificationCourseId,
+        pixScore: privateCertificateData.pixScore,
+        status: 'validated',
+      });
+
+      await databaseBuilder.commit();
+
+      // when
+      const result = await certificateRepository.findPrivateCertificatesByUserId({ userId });
+
+      // then
+      expect(result).to.have.length(1);
+    });
+
+    it('should return the certificate when it is not published', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser().id;
+      const privateCertificateData = {
+        firstName: 'Sarah Michelle',
+        lastName: 'Gellar',
+        birthdate: '1977-04-14',
+        birthplace: 'Saint-Ouen',
+        isPublished: false,
+        isCancelled: false,
+        userId,
+        date: new Date('2020-01-01'),
+        verificationCode: 'P-SOMECODE',
+        maxReachableLevelOnCertificationDate: 5,
+        deliveredAt: new Date('2021-05-05'),
+        certificationCenter: 'Centre des poules bien dodues',
+        pixScore: 51,
+      };
+      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+      const sessionId = databaseBuilder.factory.buildSession({
+        publishedAt: privateCertificateData.deliveredAt,
+        certificationCenter: privateCertificateData.certificationCenter,
+        certificationCenterId,
+      }).id;
+      const certificationCourseId = databaseBuilder.factory.buildCertificationCourse({
+        firstName: privateCertificateData.firstName,
+        lastName: privateCertificateData.lastName,
+        birthdate: privateCertificateData.birthdate,
+        birthplace: privateCertificateData.birthplace,
+        isPublished: privateCertificateData.isPublished,
+        isCancelled: privateCertificateData.isCancelled,
+        createdAt: privateCertificateData.date,
+        verificationCode: privateCertificateData.verificationCode,
+        maxReachableLevelOnCertificationDate: privateCertificateData.maxReachableLevelOnCertificationDate,
+        sessionId,
+        userId,
+      }).id;
+      databaseBuilder.factory.buildAssessmentResult.last({
+        certificationCourseId,
+        pixScore: privateCertificateData.pixScore,
+        status: 'validated',
+      });
+
+      await databaseBuilder.commit();
+
+      // when
+      const result = await certificateRepository.findPrivateCertificatesByUserId({ userId });
+
+      // then
+      expect(result).to.have.length(1);
+    });
+
+    it('should return the certificate when it is rejected', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser().id;
+      const privateCertificateData = {
+        firstName: 'Sarah Michelle',
+        lastName: 'Gellar',
+        birthdate: '1977-04-14',
+        birthplace: 'Saint-Ouen',
+        isPublished: true,
+        isCancelled: false,
+        userId,
+        date: new Date('2020-01-01'),
+        verificationCode: 'P-SOMECODE',
+        maxReachableLevelOnCertificationDate: 5,
+        deliveredAt: new Date('2021-05-05'),
+        certificationCenter: 'Centre des poules bien dodues',
+        pixScore: 51,
+      };
+      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+      const sessionId = databaseBuilder.factory.buildSession({
+        publishedAt: privateCertificateData.deliveredAt,
+        certificationCenter: privateCertificateData.certificationCenter,
+        certificationCenterId,
+      }).id;
+      const certificationCourseId = databaseBuilder.factory.buildCertificationCourse({
+        firstName: privateCertificateData.firstName,
+        lastName: privateCertificateData.lastName,
+        birthdate: privateCertificateData.birthdate,
+        birthplace: privateCertificateData.birthplace,
+        isPublished: privateCertificateData.isPublished,
+        isCancelled: privateCertificateData.isCancelled,
+        createdAt: privateCertificateData.date,
+        verificationCode: privateCertificateData.verificationCode,
+        maxReachableLevelOnCertificationDate: privateCertificateData.maxReachableLevelOnCertificationDate,
+        sessionId,
+        userId,
+      }).id;
+
+      databaseBuilder.factory.buildAssessmentResult.last({
+        certificationCourseId,
+        pixScore: privateCertificateData.pixScore,
+        status: 'rejected',
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const result = await certificateRepository.findPrivateCertificatesByUserId({ userId });
+
+      // then
+      expect(result).to.have.length(1);
+    });
+
+    it('should return a collection of PrivateCertificate', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser().id;
+      const privateCertificateData = {
+        firstName: 'Sarah Michelle',
+        lastName: 'Gellar',
+        birthdate: '1977-04-14',
+        birthplace: 'Saint-Ouen',
+        isPublished: true,
+        isCancelled: false,
+        userId,
+        date: new Date('2020-01-01'),
+        verificationCode: 'ABCDE-F',
+        maxReachableLevelOnCertificationDate: 5,
+        deliveredAt: new Date('2021-05-05'),
+        certificationCenter: 'Centre des poules bien dodues',
+        pixScore: 51,
+        commentForCandidate: 'Il aime beaucoup les mangues, et ça se voit !',
+      };
+
+      const { certificationCourseId } = await _buildValidPrivateCertificate(privateCertificateData);
+      // when
+      const privateCertificates = await certificateRepository.findPrivateCertificatesByUserId({ userId });
+
+      // then
+      const expectedPrivateCertificate = domainBuilder.buildPrivateCertificate.validated({
+        id: certificationCourseId,
+        ...privateCertificateData,
+      });
+      expect(privateCertificates).to.have.length(1);
+      expect(privateCertificates[0]).to.deepEqualInstance(expectedPrivateCertificate);
+    });
+
+    it('should return all the certificates of the user if he has many ordered by creation date DESC', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser().id;
+      const { certificationCourseId } = await _buildValidPrivateCertificate({ userId, date: '2021-05-02' });
+      const { certificationCourseId: certificationCourseId2 } = await _buildValidPrivateCertificate({
+        userId,
+        date: '2021-06-02',
+      });
+      const { certificationCourseId: certificationCourseId3 } = await _buildValidPrivateCertificate({
+        userId,
+        date: '2021-07-02',
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const privateCertificates = await certificateRepository.findPrivateCertificatesByUserId({ userId });
+
+      // then
+      expect(privateCertificates).to.have.length(3);
+      expect(privateCertificates[0].id).to.equal(certificationCourseId3);
+      expect(privateCertificates[1].id).to.equal(certificationCourseId2);
+      expect(privateCertificates[2].id).to.equal(certificationCourseId);
+    });
+
+    it('should build from the latest assessment result if validated', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser().id;
+      const privateCertificateData = {
+        firstName: 'Sarah Michelle',
+        lastName: 'Gellar',
+        birthdate: '1977-04-14',
+        birthplace: 'Saint-Ouen',
+        isPublished: true,
+        userId,
+        date: new Date('2020-01-01'),
+        verificationCode: 'ABCDE-F',
+        maxReachableLevelOnCertificationDate: 5,
+        deliveredAt: new Date('2021-05-05'),
+        certificationCenter: 'Centre des poules bien dodues',
+        pixScore: 51,
+        commentForCandidate: 'Il aime beaucoup les mangues, et ça se voit !',
+      };
+
+      const { certificationCourseId } = await _buildValidPrivateCertificateWithSeveralResults(privateCertificateData);
+
+      // when
+      const privateCertificates = await certificateRepository.findPrivateCertificatesByUserId({ userId });
+
+      // then
+      const expectedPrivateCertificate = domainBuilder.buildPrivateCertificate.validated({
+        id: certificationCourseId,
+        ...privateCertificateData,
+      });
+      expect(privateCertificates[0]).to.deepEqualInstance(expectedPrivateCertificate);
+    });
+  });
 });
 
 function _buildIncomplete(certificationAttestationData) {
@@ -1243,6 +1546,83 @@ function _buildCancelled(certificationAttestationData) {
     pixScore: certificationAttestationData.pixScore,
     status: 'validated',
   });
+}
+
+async function _buildValidPrivateCertificate(privateCertificateData, buildCompetenceMark = true) {
+  const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+  const sessionId = databaseBuilder.factory.buildSession({
+    publishedAt: privateCertificateData.deliveredAt,
+    certificationCenter: privateCertificateData.certificationCenter,
+    certificationCenterId,
+  }).id;
+  const certificationCourseId = databaseBuilder.factory.buildCertificationCourse({
+    firstName: privateCertificateData.firstName,
+    lastName: privateCertificateData.lastName,
+    birthdate: privateCertificateData.birthdate,
+    birthplace: privateCertificateData.birthplace,
+    isPublished: privateCertificateData.isPublished,
+    isCancelled: false,
+    createdAt: privateCertificateData.date,
+    verificationCode: privateCertificateData.verificationCode,
+    maxReachableLevelOnCertificationDate: privateCertificateData.maxReachableLevelOnCertificationDate,
+    sessionId,
+    userId: privateCertificateData.userId,
+  }).id;
+  const assessmentResultId = databaseBuilder.factory.buildAssessmentResult.last({
+    certificationCourseId,
+    pixScore: privateCertificateData.pixScore,
+    status: 'validated',
+    commentForCandidate: privateCertificateData.commentForCandidate,
+    commentByAutoJury: privateCertificateData.commentByAutoJury,
+    createdAt: new Date('2021-01-01'),
+  }).id;
+
+  if (buildCompetenceMark) {
+    databaseBuilder.factory.buildCompetenceMark({
+      assessmentResultId,
+    });
+  }
+
+  await databaseBuilder.commit();
+
+  return { certificationCourseId, assessmentResultId };
+}
+
+async function _buildValidPrivateCertificateWithSeveralResults(privateCertificateData) {
+  const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+  const sessionId = databaseBuilder.factory.buildSession({
+    publishedAt: privateCertificateData.deliveredAt,
+    certificationCenter: privateCertificateData.certificationCenter,
+    certificationCenterId,
+  }).id;
+  const certificationCourseId = databaseBuilder.factory.buildCertificationCourse({
+    firstName: privateCertificateData.firstName,
+    lastName: privateCertificateData.lastName,
+    birthdate: privateCertificateData.birthdate,
+    birthplace: privateCertificateData.birthplace,
+    isPublished: privateCertificateData.isPublished,
+    isCancelled: false,
+    createdAt: privateCertificateData.date,
+    verificationCode: privateCertificateData.verificationCode,
+    maxReachableLevelOnCertificationDate: privateCertificateData.maxReachableLevelOnCertificationDate,
+    sessionId,
+    userId: privateCertificateData.userId,
+  }).id;
+  const { id: lastAssessmentResultId } = databaseBuilder.factory.buildAssessmentResult.last({
+    certificationCourseId,
+    pixScore: privateCertificateData.pixScore,
+    status: 'validated',
+    commentForCandidate: privateCertificateData.commentForCandidate,
+    createdAt: new Date('2021-03-01'),
+  });
+
+  databaseBuilder.factory.buildCompetenceMark({
+    assessmentResultId: lastAssessmentResultId,
+  });
+
+  await databaseBuilder.commit();
+
+  return { certificationCourseId };
 }
 
 function _buildValidCertificationAttestation(certificationAttestationData, buildCompetenceMark = true) {
